+++ src/pg/blockchain-api/pg-blockchain-api-store.ts
@@ -1,5 +1,6 @@
 import * as postgres from 'postgres';
 import { ENV } from '../../env';
+import { connectPostgres, PgSqlClient } from '../postgres-tools';
 
 export interface BlockchainDbSmartContract {
   contract_id: string;
@@ -18,16 +19,24 @@ export interface BlockchainDbContractLog {
  * Connects and queries the Stacks Blockchain API postgres DB.
  */
 export class PgBlockchainApiStore {
-  readonly sql: postgres.Sql<any>;
+  readonly sql: PgSqlClient;
 
-  constructor() {
-    this.sql = postgres({
-      host: ENV.BLOCKCHAIN_API_PGHOST,
-      port: ENV.BLOCKCHAIN_API_PGPORT,
-      user: ENV.BLOCKCHAIN_API_PGUSER,
-      password: ENV.BLOCKCHAIN_API_PGPASSWORD,
-      database: ENV.BLOCKCHAIN_API_PGDATABASE,
+  constructor(sql: PgSqlClient) {
+    this.sql = sql;
+  }
+
+  static async connect() {
+    const sql = await connectPostgres({
+      usageName: 'tms-blockchain-api',
+      connectionArgs: {
+        host: ENV.BLOCKCHAIN_API_PGHOST,
+        port: ENV.BLOCKCHAIN_API_PGPORT,
+        user: ENV.BLOCKCHAIN_API_PGUSER,
+        password: ENV.BLOCKCHAIN_API_PGPASSWORD,
+        database: ENV.BLOCKCHAIN_API_PGDATABASE,
+      },
     });
+    return new PgBlockchainApiStore(sql);
   }
 
   async close() {
