+++ src/pg/postgres-tools/index.ts
@@ -1,98 +1,28 @@
 import * as postgres from 'postgres';
+import { isPgConnectionError } from './errors';
 import { stopwatch, timeout } from './helpers';
+import { PG_TYPE_MAPPINGS } from './types';
 
 export type PgSqlClient = postgres.Sql<any> | postgres.TransactionSql<any>;
 
-/**
- * Checks if a given error from the pg lib is a connection error (i.e. the query is retryable).
- * If true then returns a normalized error message, otherwise returns false.
- */
-export function isPgConnectionError(error: any): string | false {
-  if (error.code === 'ECONNREFUSED') {
-    return 'Postgres connection ECONNREFUSED';
-  } else if (error.code === 'ETIMEDOUT') {
-    return 'Postgres connection ETIMEDOUT';
-  } else if (error.code === 'ENOTFOUND') {
-    return 'Postgres connection ENOTFOUND';
-  } else if (error.code === 'ECONNRESET') {
-    return 'Postgres connection ECONNRESET';
-  } else if (error.code === 'CONNECTION_CLOSED') {
-    return 'Postgres connection CONNECTION_CLOSED';
-  } else if (error.code === 'CONNECTION_ENDED') {
-    return 'Postgres connection CONNECTION_ENDED';
-  } else if (error.code === 'CONNECTION_DESTROYED') {
-    return 'Postgres connection CONNECTION_DESTROYED';
-  } else if (error.code === 'CONNECTION_CONNECT_TIMEOUT') {
-    return 'Postgres connection CONNECTION_CONNECT_TIMEOUT';
-  } else if (error.code === 'CONNECT_TIMEOUT') {
-    return 'Postgres connection CONNECT_TIMEOUT';
-  } else if (error.message) {
-    const msg = (error as Error).message.toLowerCase();
-    if (msg.includes('database system is starting up')) {
-      return 'Postgres connection failed while database system is starting up';
-    } else if (msg.includes('database system is shutting down')) {
-      return 'Postgres connection failed while database system is shutting down';
-    } else if (msg.includes('connection terminated unexpectedly')) {
-      return 'Postgres connection terminated unexpectedly';
-    } else if (msg.includes('connection terminated')) {
-      return 'Postgres connection terminated';
-    } else if (msg.includes('connection error')) {
-      return 'Postgres client has encountered a connection error and is not queryable';
-    } else if (msg.includes('terminating connection due to unexpected postmaster exit')) {
-      return 'Postgres connection terminating due to unexpected postmaster exit';
-    } else if (msg.includes('getaddrinfo eai_again')) {
-      return 'Postgres connection failed due to a DNS lookup error';
-    }
-  }
-  return false;
-}
-
-const PG_TYPE_MAPPINGS = {
-  // Make both `string` and `Buffer` be compatible with a `bytea` columns.
-  // * Buffers and strings with `0x` prefixes will be transformed to hex format (`\x`).
-  // * Other strings will be passed as-is.
-  // From postgres, all values will be returned as strings with `0x` prefix.
-  bytea: {
-    to: 17,
-    from: [17],
-    serialize: (x: any) => {
-      if (typeof x === 'string') {
-        if (/^(0x|0X)[a-fA-F0-9]*$/.test(x)) {
-          // hex string with "0x" prefix
-          if (x.length % 2 !== 0) {
-            throw new Error(`Hex string is an odd number of digits: "${x}"`);
-          }
-          return '\\x' + x.slice(2);
-        } else if (x.length === 0) {
-          return '\\x';
-        } else if (/^\\x[a-fA-F0-9]*$/.test(x)) {
-          // hex string with "\x" prefix (already encoded for postgres)
-          if (x.length % 2 !== 0) {
-            throw new Error(`Hex string is an odd number of digits: "${x}"`);
-          }
-          return x;
-        } else {
-          throw new Error(`String value for bytea column does not have 0x prefix: "${x}"`);
-        }
-      } else if (Buffer.isBuffer(x)) {
-        return '\\x' + x.toString('hex');
-      } else if (ArrayBuffer.isView(x)) {
-        return '\\x' + Buffer.from(x.buffer, x.byteOffset, x.byteLength).toString('hex');
-      } else {
-        throw new Error(
-          `Cannot serialize unexpected type "${x.constructor.name}" to bytea hex string`
-        );
-      }
-    },
-    parse: (x: any) => `0x${x.slice(2)}`,
-  },
+export type PgConnectionUri = string;
+export type PgConnectionVars = {
+  database: string;
+  user: string;
+  password: string;
+  host: string;
+  port: number;
+  schema?: string;
+  ssl?: boolean;
+  applicationName?: string;
 };
-/** Values will be automatically converted into a `bytea` compatible string before sending to pg. */
-export type PgBytea = string | Buffer;
-/** The `string` type guarantees the value will fit into the `numeric` pg type. */
-export type PgNumeric = string;
-/** JSON objects will be automatically stringified before insertion. */
-export type PgJsonb = any;
+export type PgConnectionArgs = PgConnectionUri | PgConnectionVars;
+
+export type PgConnectionConfig = {
+  idleTimeout?: number;
+  maxLifetime?: number;
+  poolMax?: number;
+};
 
 /**
  * Connects to Postgres. This function will also test the connection first to make sure
@@ -102,10 +32,12 @@ export type PgJsonb = any;
  */
 export async function connectPostgres({
   usageName,
-  pgServer,
+  connectionArgs,
+  connectionConfig,
 }: {
   usageName: string;
-  pgServer: PgServer;
+  connectionArgs: PgConnectionArgs;
+  connectionConfig?: PgConnectionConfig;
 }): Promise<PgSqlClient> {
   const initTimer = stopwatch();
   let connectionError: Error | undefined;
@@ -114,14 +46,15 @@ export async function connectPostgres({
   do {
     const testSql = getPostgres({
       usageName: `${usageName};conn-poll`,
-      pgServer: pgServer,
+      connectionArgs,
+      connectionConfig,
     });
     try {
       await testSql`SELECT version()`;
       connectionOkay = true;
       break;
     } catch (error: any) {
-      if (isPgConnectionError(error) || error instanceof postgres.PostgresError) {
+      if (isPgConnectionError(error)) {
         const timeElapsed = initTimer.getElapsed();
         if (timeElapsed - lastElapsedLog > 2000) {
           lastElapsedLog = timeElapsed;
@@ -143,42 +76,25 @@ export async function connectPostgres({
   }
   const sql = getPostgres({
     usageName: `${usageName};datastore-crud`,
-    pgServer: pgServer,
+    connectionArgs,
+    connectionConfig,
   });
   return sql;
 }
 
 export function getPostgres({
   usageName,
-  pgServer,
+  connectionArgs,
+  connectionConfig,
 }: {
   usageName: string;
-  pgServer?: PgServer;
+  connectionArgs: PgConnectionArgs;
+  connectionConfig?: PgConnectionConfig;
 }): PgSqlClient {
-  const pgEnvVars = {
-    database: getPgConnectionEnvValue('DATABASE', pgServer),
-    user: getPgConnectionEnvValue('USER', pgServer),
-    password: getPgConnectionEnvValue('PASSWORD', pgServer),
-    host: getPgConnectionEnvValue('HOST', pgServer),
-    port: getPgConnectionEnvValue('PORT', pgServer),
-    ssl: getPgConnectionEnvValue('SSL', pgServer),
-    schema: getPgConnectionEnvValue('SCHEMA', pgServer),
-    applicationName: getPgConnectionEnvValue('APPLICATION_NAME', pgServer),
-    idleTimeout: parseInt(getPgConnectionEnvValue('IDLE_TIMEOUT', pgServer) ?? '30'),
-    maxLifetime: parseInt(getPgConnectionEnvValue('MAX_LIFETIME', pgServer) ?? '60'),
-    poolMax: parseInt(process.env['PG_CONNECTION_POOL_MAX'] ?? '10'),
-  };
-  const defaultAppName = 'stacks-blockchain-api';
-  const pgConnectionUri = getPgConnectionEnvValue('CONNECTION_URI', pgServer);
-  const pgConfigEnvVar = Object.entries(pgEnvVars).find(([, v]) => typeof v === 'string')?.[0];
-  if (pgConfigEnvVar && pgConnectionUri) {
-    throw new Error(
-      `Both PG_CONNECTION_URI and ${pgConfigEnvVar} environmental variables are defined. PG_CONNECTION_URI must be defined without others or omitted.`
-    );
-  }
+  const defaultAppName = 'postgres';
   let sql: PgSqlClient;
-  if (pgConnectionUri) {
-    const uri = new URL(pgConnectionUri);
+  if (typeof connectionArgs === 'string') {
+    const uri = new URL(connectionArgs);
     const searchParams = Object.fromEntries(
       [...uri.searchParams.entries()].map(([k, v]) => [k.toLowerCase(), v])
     );
@@ -193,28 +109,28 @@ export function getPostgres({
     uri.searchParams.set('application_name', appName);
     sql = postgres(uri.toString(), {
       types: PG_TYPE_MAPPINGS,
-      max: pgEnvVars.poolMax,
+      max: connectionConfig?.poolMax ?? 10,
       connection: {
         application_name: appName,
         search_path: schema,
       },
     });
   } else {
-    const appName = `${pgEnvVars.applicationName ?? defaultAppName}:${usageName}`;
+    const appName = `${connectionArgs.applicationName ?? defaultAppName}:${usageName}`;
     sql = postgres({
-      database: pgEnvVars.database,
-      user: pgEnvVars.user,
-      password: pgEnvVars.password,
-      host: pgEnvVars.host,
-      port: parsePort(pgEnvVars.port),
-      ssl: parseArgBoolean(pgEnvVars.ssl),
-      idle_timeout: pgEnvVars.idleTimeout,
-      max_lifetime: pgEnvVars.maxLifetime,
-      max: pgEnvVars.poolMax,
+      database: connectionArgs.database,
+      user: connectionArgs.user,
+      password: connectionArgs.password,
+      host: connectionArgs.host,
+      port: connectionArgs.port,
+      ssl: connectionArgs.ssl,
+      idle_timeout: connectionConfig?.idleTimeout,
+      max_lifetime: connectionConfig?.maxLifetime,
+      max: connectionConfig?.poolMax,
       types: PG_TYPE_MAPPINGS,
       connection: {
         application_name: appName,
-        search_path: pgEnvVars.schema,
+        search_path: connectionArgs.schema,
       },
     });
   }
