+++ tests/process-smart-contract-job.test.ts
@@ -11,7 +11,7 @@ describe('ProcessSmartContractJob', () => {
 
   beforeEach(async () => {
     ENV.PGDATABASE = 'postgres';
-    db = new PgStore();
+    db = await PgStore.connect();
     await cycleMigrations();
   });
